refactor(settings): extract formatDateForInput and simplify masked input
- Extract shared formatDateForInput to settings-constants to eliminate duplication
- Simplify MaskedAadhaarInput by removing complex onChange handler (onKeyDown handles all input)
- Export new components from settings barrel file
