+++ src/components/settings/index.ts
@@ -9,4 +9,5 @@ export { NotificationsSection } from './notifications-section'
 export { NotificationChannelGroup } from './notification-channel-group'
 export { CHANNEL_ITEMS, CATEGORY_ITEMS } from './notification-constants'
 export { AccountSection } from './account-section'
-export { BLOOD_GROUP_OPTIONS, GENDER_OPTIONS } from './settings-constants'
+export { BLOOD_GROUP_OPTIONS, GENDER_OPTIONS, formatDateForInput } from './settings-constants'
+export { MaskedAadhaarInput, maskAadhaarDisplay } from './masked-aadhaar-input'
