+++ src/components/settings/settings-constants.ts
@@ -1,5 +1,14 @@
 import type { BloodGroup, Gender } from '@/types/profile'
 
+/** Converts an ISO date string to "YYYY-MM-DD" format for HTML date inputs. */
+export function formatDateForInput(dateStr: string): string {
+  const d = new Date(dateStr)
+  const year = d.getUTCFullYear()
+  const month = String(d.getUTCMonth() + 1).padStart(2, '0')
+  const day = String(d.getUTCDate()).padStart(2, '0')
+  return `${year}-${month}-${day}`
+}
+
 export const BLOOD_GROUP_OPTIONS: { value: BloodGroup; label: string }[] = [
   { value: 'A+', label: 'A+' },
   { value: 'A-', label: 'A-' },
